Add language switcher to member area
