+++ apps/store/src/blocks/FooterBlock/LanguageSelector.tsx
@@ -1,7 +1,6 @@
 'use client'
 import { useTranslation } from 'next-i18next'
 import { type ChangeEventHandler } from 'react'
-import { languageSelectForm } from '@/blocks/FooterBlock/LanguageSelectForm.css'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
 import { getCountryLocale } from '@/utils/l10n/countryUtils'
 import { LocaleField } from '@/utils/l10n/locales'
@@ -11,7 +10,7 @@ import { useChangeLocale } from '@/utils/l10n/useChangeLocale'
 import { useCurrentCountry } from '@/utils/l10n/useCurrentCountry'
 import { useCurrentLocale } from '@/utils/l10n/useCurrentLocale'
 
-export function LanguageSelectForm() {
+export function LanguageSelector() {
   const currentCountry = useCurrentCountry()
   const { t } = useTranslation()
   const handleLocaleChange = useChangeLocale()
@@ -32,14 +31,13 @@ export function LanguageSelectForm() {
   const { language: currentLanguage } = useCurrentLocale()
 
   return (
-    <div className={languageSelectForm}>
-      <InputSelect
-        name={LocaleField.Language}
-        onChange={handleChangeLanguage}
-        defaultValue={currentLanguage}
-        options={languageList}
-        backgroundColor="gray300"
-      />
-    </div>
+    <InputSelect
+      name={LocaleField.Language}
+      onChange={handleChangeLanguage}
+      defaultValue={currentLanguage}
+      options={languageList}
+      backgroundColor="gray300"
+      size="small"
+    />
   )
 }
